# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PaginationResponse_ProcessRepresentationResponse_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProcessRepresentationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginationResponse[ProcessRepresentationResponse]",
        "type": "object"
      },
      "PaginationResponse_ProcessSummaryResponse_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProcessSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginationResponse[ProcessSummaryResponse]",
        "type": "object"
      },
...
      "ProcessDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_collection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestCollectionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Observed At"
          },
          "numero_cnj": {
            "pattern": "^\\d{20}$",
            "title": "Numero Cnj",
            "type": "string"
          },
          "representation_count": {
            "minimum": 0.0,
            "title": "Representation Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "numero_cnj",
          "created_at",
          "representation_count",
          "latest_observed_at",
          "latest_collection"
        ],
        "title": "ProcessDetailResponse",
        "type": "object"
      },
...
      "ProcessMovementsResponse": {
        "additionalProperties": false,
        "properties": {
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/MovementDiagnosticResponse"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProcessMovementResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "process_id": {
            "format": "uuid",
            "title": "Process Id",
            "type": "string"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "process_id",
          "diagnostics"
        ],
        "title": "ProcessMovementsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API local de jobs, processos, representações, movimentos e presets.",
    "title": "AgroJud Radar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/processes": {
      "get": {
        "operationId": "list_processes_api_v1_processes_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "process_number",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 30,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Process Number"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "in": "query",
            "name": "class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class"
            }
          },
          {
            "in": "query",
            "name": "court_unit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Court Unit"
            }
          },
          {
            "in": "query",
            "name": "collection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collection Id"
            }
          },
          {
            "in": "query",
            "name": "preset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse_ProcessSummaryResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Pesquisa processos persistidos",
        "tags": [
          "processes"
        ]
      }
    },
    "/api/v1/processes/{process_id}": {
      "get": {
        "operationId": "get_process_api_v1_processes__process_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Process Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Consulta um processo local",
        "tags": [
          "processes"
        ]
      }
    },
    "/api/v1/processes/{process_id}/movements": {
      "get": {
        "operationId": "list_movements_api_v1_processes__process_id__movements_get",
        "parameters": [
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Process Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessMovementsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Lista movimentos normalizados sem misturar representações",
        "tags": [
          "processes"
        ]
      }
    },
    "/api/v1/processes/{process_id}/representations": {
      "get": {
        "operationId": "list_representations_api_v1_processes__process_id__representations_get",
        "parameters": [
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Process Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse_ProcessRepresentationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Lista representações e seus diagnósticos",
        "tags": [
          "processes"
        ]
      }
    }
  }
}